+++ Assets/_Scripts/Player/Experience/HealthOrb.cs
@@ -3,12 +3,12 @@ using UnityEngine;
 public class HealthOrb : OrbBase
 {
     [Header("Configuración de Vida")]
-    [SerializeField] private float minHeal = 10f;
-    [SerializeField] private float maxHeal = 25f;
+    [SerializeField] private int minHeal = 10;
+    [SerializeField] private int maxHeal = 25;
 
     protected override void ApplyEffect(GameObject player)
     {
-        float healAmount = Random.Range(minHeal, maxHeal + 1);
+        int healAmount = Random.Range(minHeal, maxHeal + 1);
 
         if (player.TryGetComponent(out PlayerContext context))
         {
